+++ cli/repl/commands/describe_last_command.h
@@ -1,0 +1,9 @@
+#pragma once
+
+#include "registry.h"
+
+namespace xsql::cli {
+
+CommandHandler make_describe_last_command();
+
+}  // namespace xsql::cli